netlib: fix nxstyle issue for netlink related code

+++ netutils/netlib/netlib_getroute.c
@@ -161,7 +161,7 @@ ssize_t netlib_get_route(FAR struct rtentry *rtelist,
   addr.nl_groups = 0;
 
   ret = bind(fd, (FAR const struct sockaddr *)&addr,
-             sizeof( struct sockaddr_nl));
+             sizeof(struct sockaddr_nl));
   if (ret < 0)
     {
       int errcode = errno;
@@ -258,7 +258,8 @@ ssize_t netlib_get_route(FAR struct rtentry *rtelist,
               if (nrecvd <= sizeof(struct netlib_recvfrom_response_s) ||
                   u.resp.rte.rtm_family != family)
                 {
-                  fprintf(stderr, "ERROR: Bad massage size: %ld\n", (long)nrecvd);
+                  fprintf(stderr, "ERROR: Bad massage size: %ld\n",
+                          (long)nrecvd);
                   goto errout_with_socket;
                 }
 
@@ -320,8 +321,8 @@ ssize_t netlib_get_route(FAR struct rtentry *rtelist,
             break;
 
           default:
-            fprintf(stderr, "ERROR: Message type %u, length %lu\n",
-                    u.resp.hdr.nlmsg_type, (unsigned long)u.resp.hdr.nlmsg_len);
+            fprintf(stderr, "ERROR: Message type %u, length %u\n",
+                    u.resp.hdr.nlmsg_type, (unsigned)u.resp.hdr.nlmsg_len);
             ret = -EIO;
             goto errout_with_socket;
         }
